+++ CHANGELOG.md
@@ -1,18 +1,20 @@
 ## [%new_version] - %date
 
 ### Added
-- 
 - A prompt at the end of `tui` configuration for a new project whether to open `projects.json` in a text editor.
+- `global_exception_handler` to handle any uncaught exceptions, generate a report and create an issue archive.
 
 ### Changed
 - 
 - The way `Packer` class checks modified, deleted and added files to use Git instead of utilizing the previous and new integrity.
 - `revert_changes` function, starts with `print_and_log`, since every use case was already manually doing it every time and also added to a call in case git didn't successfully commit.
-- All user tied paths to be handled in `config.py` and then just be imported by all parts of the program that needs them.
+- All user tied paths to be handled in `paths.py` and then just be imported by all parts of the program that needs them.
+- Packer version to be stored in `config.py` and them imported by everyone.
 
 ### Fixed
 - Git branch logic to update the origin (Github's) version of those branches.
 - New version git tag landing on the same as the previous tag, now the tag is explicit attached to the latest commit vai sha.
+- Packer version to be a different version identifier than the project being updated.
 
 ---